Provide integration tests via the key Ditto features - ContainerFactory
[#70] Provide integration tests via the key Ditto features - ContainerFactory
- refactor the code to use the common logic
